WIP for #13, see TODO below
Not only is the event matching logic getting sort of complicated, I
noticed a logical inconsistency between the way I want to implement
comment author matching and the way we are implementing change kind
matching.

My use case for comment author matching is to filter (match on a null
channel) for a specific author. My use case for change kind is to
filter all *except* a specific change kind.

The logical inconsistency is currently the fact that IF the comment
author property is present, I want to match ONLY comment added events
with a matching author, and have the matcher fail everything else. In
constrast, for change kind, if the property exists I match patch set
created events with a matching change kind, as well as all events that
aren't patch set created events.

If comment author matching worked the same way, a rule to filter out
all comments added by user "gerrit" would filter them, as well as
every other event that wasn't a comment added event, which is clearly
broken.

I need to make sure the matching rules work consistently. What sucks
about that is I'd need to take the current "filter everything but
REWORK" and replace it with individual rules for every single type of
change kind except that. It requires me to know them all, which sucks.
Maybe another approach would be to make change kind and authors and
the like pattern match on regex. It still requires me to know them
all. Unless I can do a negative lookahead assertion in Java, not sure.

Overall I need to give this more thought and have a better set of
tests. I need a more integration style test that uses a set of
matchers and asserts about the whole group, because I'm worried about
matcher interactions.

Also, I'd like to look into some way to have the matching done more
flexibly... like some sort of runtime compilation or extension
language. Or
https://discuss.kotlinlang.org/t/loading-source-at-runtime/58/3